hms-enterprise-grade/services/audit/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
SQLAlchemy==2.0.36
python-jose[cryptography]==3.3.0
prometheus-fastapi-instrumentator==7.0.0
psycopg2-binary==2.9.10
fastapi-limiter==0.1.6
redis==5.0.8
cryptography==43.0.3
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
